Dan #atomic #refac Replicate update files testing into content_spec
(from update controller spec: moving responsibility of files handling
away from controllers to content object)
